requirements.txt
Flask
sqlalchemy
PyMySQL
PyMySQL[rsa]
python-dotenv